v. 0.2
There were potential issues with breaking the Soul Gem, and it was just too cheap to do so. This also changed the number of break counters from 3 -> 5, so that a player can't break it in a single turn. The revised price based on current pieces in the Soul Gem incentivizes summoning pieces before attempting a break, otherwise there's no reason not to simply attempt a break from move one.